refactor: 배송비 반환 타입 long 통일 및 자동 마감 폴링 로그 보강
- ShippingFeePolicy.feeFor·Buncheol.shippingFeeFor 반환 int→long (bidAmount·paymentAmount 등 금액 필드와 타입 일관화)
- 자동 마감 폴링 완료 로그에 실패 건수(failedCount) 추가로 관측성 개선

+++ src/main/java/buncheoleasy/buncheol/domain/Buncheol.java
@@ -112,7 +112,7 @@ public class Buncheol extends TimestampedEntity implements Cursorable {
   }
 
   // 참여자가 선택한 배송수단의 배송비. 낙찰자 결제 금액(제시가 + 배송비) 계산에 사용한다.
-  public int shippingFeeFor(final ShippingMethod shippingMethod) {
+  public long shippingFeeFor(final ShippingMethod shippingMethod) {
     return shippingFeePolicy.feeFor(shippingMethod);
   }
